Migrate implicit self, constant, parameter, and definition tests from Rust to Ruby
The Rust analyzer tests for these domains were full-pipeline integration
tests (parse → install → propagate → resolve) that duplicate the Ruby
integration tests. Move unique scenarios to Ruby and remove redundant
Rust tests, keeping only true unit tests that verify internal state
not observable through the CLI.

- definitions.rs: 20 → 12 tests (remove 8 redundant)
- dispatch.rs: remove implicit self, constant read, keyword arg tests
- parameters.rs: 11 → 3 tests (remove 8 redundant)
- Add 2 implicit self call tests to Ruby (module, block)
- Add 1 constant read test to Ruby (deeply nested)
- Add 1 parameter test to Ruby (keyword arg via .new)
