+++ core/src/analyzer/definitions.rs
@@ -348,29 +348,6 @@ mod tests {
     }
 
     #[test]
-    fn test_process_def_node_registers_user_method() {
-        let source = "class User; def name; \"Alice\"; end; end";
-        let session = ParseSession::new();
-        let parse_result = session.parse_source(source, "test.rb").unwrap();
-        let root = parse_result.node();
-        let program = root.as_program_node().unwrap();
-
-        let mut genv = GlobalEnv::new();
-        let mut lenv = LocalEnv::new();
-        let mut changes = ChangeSet::new();
-
-        let stmt = program.statements().body().first().unwrap();
-        let class_node = stmt.as_class_node().unwrap();
-        process_class_node(&mut genv, &mut lenv, &mut changes, source, &class_node);
-
-        // User#name should be registered as a user-defined method
-        let info = genv
-            .resolve_method(&Type::instance("User"), "name")
-            .expect("User#name should be registered");
-        assert!(info.return_vertex.is_some());
-    }
-
-    #[test]
     fn test_qualified_name_method_registration() {
         let source = r#"
 module Api
@@ -407,144 +384,6 @@ end
             genv.resolve_method(&Type::instance("User"), "name").is_none(),
             "User#name should not exist — method should be registered under qualified name"
         );
-    }
-
-    #[test]
-    fn test_same_class_name_different_namespace() {
-        let source = r#"
-module Api
-  class User
-    def name
-      "api_user"
-    end
-  end
-end
-
-module Admin
-  class User
-    def name
-      "admin_user"
-    end
-  end
-end
-"#;
-        let session = ParseSession::new();
-        let parse_result = session.parse_source(source, "test.rb").unwrap();
-        let root = parse_result.node();
-        let program = root.as_program_node().unwrap();
-
-        let mut genv = GlobalEnv::new();
-        let mut lenv = LocalEnv::new();
-        let mut changes = ChangeSet::new();
-
-        for stmt in &program.statements().body() {
-            crate::analyzer::install::install_node(&mut genv, &mut lenv, &mut changes, source, &stmt);
-        }
-
-        // Both should be registered separately
-        let api_info = genv
-            .resolve_method(&Type::instance("Api::User"), "name")
-            .expect("Api::User#name should be registered");
-        assert!(api_info.return_vertex.is_some());
-
-        let admin_info = genv
-            .resolve_method(&Type::instance("Admin::User"), "name")
-            .expect("Admin::User#name should be registered");
-        assert!(admin_info.return_vertex.is_some());
-
-        // Simple "User" should not resolve
-        assert!(
-            genv.resolve_method(&Type::instance("User"), "name").is_none(),
-            "User#name should not exist — both are under qualified names"
-        );
-    }
-
-    #[test]
-    fn test_class_method_registration() {
-        let source = r#"
-class User
-  def self.create
-    "created"
-  end
-end
-"#;
-        let session = ParseSession::new();
-        let parse_result = session.parse_source(source, "test.rb").unwrap();
-        let root = parse_result.node();
-        let program = root.as_program_node().unwrap();
-
-        let mut genv = GlobalEnv::new();
-        let mut lenv = LocalEnv::new();
-        let mut changes = ChangeSet::new();
-
-        for stmt in &program.statements().body() {
-            crate::analyzer::install::install_node(&mut genv, &mut lenv, &mut changes, source, &stmt);
-        }
-
-        // def self.create should be registered as singleton method
-        let info = genv
-            .resolve_method(&Type::singleton("User"), "create")
-            .expect("User.create should be registered as singleton method");
-        assert!(info.return_vertex.is_some());
-    }
-
-    #[test]
-    fn test_class_method_with_params() {
-        let source = r#"
-class User
-  def self.find(id)
-    "user"
-  end
-end
-"#;
-        let session = ParseSession::new();
-        let parse_result = session.parse_source(source, "test.rb").unwrap();
-        let root = parse_result.node();
-        let program = root.as_program_node().unwrap();
-
-        let mut genv = GlobalEnv::new();
-        let mut lenv = LocalEnv::new();
-        let mut changes = ChangeSet::new();
-
-        for stmt in &program.statements().body() {
-            crate::analyzer::install::install_node(&mut genv, &mut lenv, &mut changes, source, &stmt);
-        }
-
-        let info = genv
-            .resolve_method(&Type::singleton("User"), "find")
-            .expect("User.find should be registered");
-        assert!(info.return_vertex.is_some());
-        assert_eq!(info.param_vertices.as_ref().unwrap().len(), 1);
-    }
-
-    #[test]
-    fn test_class_method_in_qualified_namespace() {
-        let source = r#"
-module Api
-  class User
-    def self.create
-      "created"
-    end
-  end
-end
-"#;
-        let session = ParseSession::new();
-        let parse_result = session.parse_source(source, "test.rb").unwrap();
-        let root = parse_result.node();
-        let program = root.as_program_node().unwrap();
-
-        let mut genv = GlobalEnv::new();
-        let mut lenv = LocalEnv::new();
-        let mut changes = ChangeSet::new();
-
-        for stmt in &program.statements().body() {
-            crate::analyzer::install::install_node(&mut genv, &mut lenv, &mut changes, source, &stmt);
-        }
-
-        let info = genv
-            .resolve_method(&Type::singleton("Api::User"), "create")
-            .expect("Api::User.create should be registered");
-        assert!(info.return_vertex.is_some());
     }
 
     #[test]
@@ -603,129 +442,5 @@ end
             genv.resolve_method(&Type::singleton("User"), "foo").is_none(),
             "User.foo should not exist — receiver is not self"
         );
-    }
-
-    #[test]
-    fn test_class_method_return_type_inference() {
-        let source = r#"
-class User
-  def self.create
-    "created"
-  end
-end
-"#;
-        let session = ParseSession::new();
-        let parse_result = session.parse_source(source, "test.rb").unwrap();
-        let root = parse_result.node();
-        let program = root.as_program_node().unwrap();
-
-        let mut genv = GlobalEnv::new();
-        let mut lenv = LocalEnv::new();
-        let mut changes = ChangeSet::new();
-
-        for stmt in &program.statements().body() {
-            crate::analyzer::install::install_node(&mut genv, &mut lenv, &mut changes, source, &stmt);
-        }
-
-        let info = genv
-            .resolve_method(&Type::singleton("User"), "create")
-            .expect("User.create should be registered");
-        let ret_vtx = info.return_vertex.expect("should have return vertex");
-
-        // Run solver to propagate types
-        genv.apply_changes(changes);
-        genv.run_all();
-
-        let vertex = genv.get_vertex(ret_vtx).or_else(|| {
-            // return vertex might be a source
-            None
-        });
-        if let Some(v) = vertex {
-            assert_eq!(v.show(), "String");
-        } else {
-            // Check if it's a source
-            let src = genv.get_source(ret_vtx).expect("should have source or vertex");
-            assert_eq!(src.ty, Type::string());
-        }
-    }
-
-    #[test]
-    fn test_class_method_in_reopened_class() {
-        let source = r#"
-class User
-  def self.create
-    "created"
-  end
-end
-
-class User
-  def self.destroy
-    "destroyed"
-  end
-end
-"#;
-        let session = ParseSession::new();
-        let parse_result = session.parse_source(source, "test.rb").unwrap();
-        let root = parse_result.node();
-        let program = root.as_program_node().unwrap();
-
-        let mut genv = GlobalEnv::new();
-        let mut lenv = LocalEnv::new();
-        let mut changes = ChangeSet::new();
-
-        for stmt in &program.statements().body() {
-            crate::analyzer::install::install_node(&mut genv, &mut lenv, &mut changes, source, &stmt);
-        }
-
-        // Both class methods should be registered
-        assert!(
-            genv.resolve_method(&Type::singleton("User"), "create").is_some(),
-            "User.create should be registered"
-        );
-        assert!(
-            genv.resolve_method(&Type::singleton("User"), "destroy").is_some(),
-            "User.destroy should be registered"
-        );
-    }
-
-    #[test]
-    fn test_class_method_param_type_propagation() {
-        let source = r#"
-class User
-  def self.find(id)
-    id
-  end
-end
-
-User.find(42)
-"#;
-        let session = ParseSession::new();
-        let parse_result = session.parse_source(source, "test.rb").unwrap();
-        let root = parse_result.node();
-        let program = root.as_program_node().unwrap();
-
-        let mut genv = GlobalEnv::new();
-        let mut lenv = LocalEnv::new();
-        let mut changes = ChangeSet::new();
-
-        for stmt in &program.statements().body() {
-            crate::analyzer::install::install_node(&mut genv, &mut lenv, &mut changes, source, &stmt);
-        }
-
-        let info = genv
-            .resolve_method(&Type::singleton("User"), "find")
-            .expect("User.find should be registered");
-        let param_vtxs = info.param_vertices.as_ref().expect("should have param vertices");
-        assert_eq!(param_vtxs.len(), 1);
-
-        let param_vtx = param_vtxs[0];
-
-        // Run solver to propagate argument types
-        genv.apply_changes(changes);
-        genv.run_all();
-
-        // Parameter should have Integer type propagated from call site
-        let vertex = genv.get_vertex(param_vtx).expect("param vertex should exist");
-        assert_eq!(vertex.show(), "Integer");
     }
 }
