+++ core/src/analyzer/parameters.rs
@@ -208,36 +208,6 @@ pub(crate) fn install_parameters(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::analyzer::install::AstInstaller;
-    use crate::parser::ParseSession;
-
-    fn analyze(source: &str) -> GlobalEnv {
-        let session = ParseSession::new();
-        let parse_result = session.parse_source(source, "test.rb").unwrap();
-        let root = parse_result.node();
-        let program = root.as_program_node().unwrap();
-
-        let mut genv = GlobalEnv::new();
-        let mut lenv = LocalEnv::new();
-
-        let mut installer = AstInstaller::new(&mut genv, &mut lenv, source);
-        for stmt in &program.statements().body() {
-            installer.install_node(&stmt);
-        }
-        installer.finish();
-
-        genv
-    }
-
-    fn get_type_show(genv: &GlobalEnv, vtx: VertexId) -> String {
-        if let Some(vertex) = genv.get_vertex(vtx) {
-            vertex.show()
-        } else if let Some(source) = genv.get_source(vtx) {
-            source.ty.show()
-        } else {
-            panic!("vertex {:?} not found as either Vertex or Source", vtx);
-        }
-    }
 
     #[test]
     fn test_install_required_parameter() {
@@ -295,135 +265,5 @@ mod tests {
         // Type should propagate from default value
         let vertex = genv.get_vertex(vtx).unwrap();
         assert_eq!(vertex.show(), "Integer");
-    }
-
-    #[test]
-    fn test_required_parameter_type_propagation() {
-        let source = r#"
-class Foo
-  def greet(name)
-    name
-  end
-end
-
-Foo.new.greet("Alice")
-"#;
-        let genv = analyze(source);
-        let info = genv.resolve_method(&Type::instance("Foo"), "greet").unwrap();
-        let ret_vtx = info.return_vertex.unwrap();
-        assert_eq!(get_type_show(&genv, ret_vtx), "String");
-    }
-
-    #[test]
-    fn test_optional_parameter_default_type() {
-        let source = r#"
-class Foo
-  def greet(name = "World")
-    name
-  end
-end
-"#;
-        let genv = analyze(source);
-        let info = genv.resolve_method(&Type::instance("Foo"), "greet").unwrap();
-        let ret_vtx = info.return_vertex.unwrap();
-        assert_eq!(get_type_show(&genv, ret_vtx), "String");
-    }
-
-    #[test]
-    fn test_multiple_parameters_from_call_site() {
-        let source = r#"
-class Calc
-  def add(x, y)
-    x
-  end
-end
-
-Calc.new.add(1, 2)
-"#;
-        let genv = analyze(source);
-        let info = genv.resolve_method(&Type::instance("Calc"), "add").unwrap();
-        let param_vtxs = info.param_vertices.as_ref().unwrap();
-        assert_eq!(param_vtxs.len(), 2);
-        // Verify return type is Integer (method returns x, which receives 1)
-        let ret_vtx = info.return_vertex.unwrap();
-        assert_eq!(get_type_show(&genv, ret_vtx), "Integer");
-    }
-
-    #[test]
-    fn test_keyword_parameter_propagation() {
-        let source = r#"
-class Foo
-  def greet(name:)
-    name
-  end
-end
-
-Foo.new.greet(name: "Alice")
-"#;
-        let genv = analyze(source);
-        let info = genv.resolve_method(&Type::instance("Foo"), "greet").unwrap();
-        let ret_vtx = info.return_vertex.unwrap();
-        assert_eq!(get_type_show(&genv, ret_vtx), "String");
-    }
-
-    #[test]
-    fn test_optional_keyword_parameter_default() {
-        let source = r#"
-class Counter
-  def count(step: 1)
-    step
-  end
-end
-"#;
-        let genv = analyze(source);
-        let info = genv.resolve_method(&Type::instance("Counter"), "count").unwrap();
-        let ret_vtx = info.return_vertex.unwrap();
-        assert_eq!(get_type_show(&genv, ret_vtx), "Integer");
-    }
-
-    #[test]
-    fn test_mixed_positional_and_keyword_params() {
-        let source = r#"
-class User
-  def initialize(id, name:)
-    @id = id
-    @name = name
-  end
-end
-
-User.new(1, name: "Alice")
-"#;
-        let genv = analyze(source);
-        assert!(genv.type_errors.is_empty());
-    }
-
-    #[test]
-    fn test_rest_parameter() {
-        let source = r#"
-class Foo
-  def bar(*args)
-    args
-  end
-end
-"#;
-        let genv = analyze(source);
-        let info = genv.resolve_method(&Type::instance("Foo"), "bar").unwrap();
-        let ret_vtx = info.return_vertex.unwrap();
-        assert_eq!(get_type_show(&genv, ret_vtx), "Array");
-    }
-
-    #[test]
-    fn test_no_parameters() {
-        let source = r#"
-class Foo
-  def bar
-    "hello"
-  end
-end
-"#;
-        let genv = analyze(source);
-        let info = genv.resolve_method(&Type::instance("Foo"), "bar").unwrap();
-        let param_vtxs = info.param_vertices.as_ref().unwrap();
-        assert!(param_vtxs.is_empty());
     }
 }
